+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### Changed
 - Database image
 
+### Removed
+- Stats from database image for soon sql database integration
+
 ## [Alpha 1.3.4](https://github.com/Nevah5/wordlebot/releases/tag/a1.3.4) - 07.02.2022
 ### Changed
 - Intents